requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.2
executing==1.2.0
Flask==2.3.2
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
idna==3.4
ipdb==0.13.13
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
line-bot-sdk==3.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
